+++ IspDataDb/dbo/Tables/DistributionScheme.sql
@@ -1,5 +1,5 @@
 ﻿CREATE TABLE [dbo].[DistributionScheme] (
-    [IdNo]                      INT            IDENTITY (1, 1) NOT NULL,
+    [IdNo]                      SMALLINT       IDENTITY (1, 1) NOT NULL,
     [DistributionSchemeCode]    VARCHAR (5)    COLLATE SQL_Latin1_General_CP1_CI_AS NOT NULL,
     [DistributionSchemeName]    VARCHAR (50)   COLLATE SQL_Latin1_General_CP1_CI_AS NOT NULL,
     [DistributionSchemeNameAra] NVARCHAR (50)  COLLATE SQL_Latin1_General_CP1_CI_AS NOT NULL,
@@ -9,6 +9,8 @@
     [DateTimeStamp]             ROWVERSION     NOT NULL,
     CONSTRAINT [PK_DistributionScheme] PRIMARY KEY CLUSTERED ([IdNo] ASC)
 );
+
+
 
 
 GO
